requirements.txt
requests==2.32.3
snowflake-connector-python==3.12.4
pandas==2.2.3
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
